// apps/admin/src/lib/api/generated/models/pageTopicProposalOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * DevFeed Admin API
 * Private administration API. OIDC sessions and CSRF protection required.
 * OpenAPI spec version: 0.0.3
 */
import type { TopicProposalOut } from './topicProposalOut';

export interface PageTopicProposalOut {
  items: TopicProposalOut[];
  limit: number;
  offset: number;
  total: number;
}